[REFACTOR] #71 - RandomPhotoColors를 Color 파일 위치로 이동
